+++ packages/exampleReactNative/babel.config.js
@@ -1,6 +1,0 @@
-// Copyright 2017-2021 @polkadot/example-reactnative authors & contributors
-// SPDX-License-Identifier: Apache-2.0
-
-module.exports = {
-  presets: ['module:metro-react-native-babel-preset']
-};