+++ packages/react-identicon/src/bundle.ts
@@ -1,0 +1,7 @@
+// Copyright 2017-2021 @polkadot/react-identicon authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export * from './icons';
+
+export { Identicon } from './Identicon';
+export { packageInfo } from './packageInfo';