+++ packages/ui-settings/src/Settings.ts
@@ -1,14 +1,14 @@
 // Copyright 2017-2021 @polkadot/ui-settings authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { Option, SettingsStruct } from './types';
+import type { Endpoint, EndpointType, Option, SettingsStruct } from './types';
 
 import EventEmitter from 'eventemitter3';
 import store from 'store';
 
-import { isUndefined } from '@polkadot/util';
+import { hasProcess, isUndefined } from '@polkadot/util';
 
-import { CAMERA, CAMERA_DEFAULT, CRYPTOS, CRYPTOS_ETH, CRYPTOS_LEDGER, ENDPOINT_DEFAULT, ENDPOINTS, ICON_DEFAULT, ICONS, LANGUAGE_DEFAULT, LEDGER_CONN, LEDGER_CONN_DEFAULT, LOCKING, LOCKING_DEFAULT, PREFIX_DEFAULT, PREFIXES, UIMODE_DEFAULT, UIMODES, UITHEME_DEFAULT, UITHEMES } from './defaults';
+import { CAMERA, CAMERA_DEFAULT, CRYPTOS, CRYPTOS_ETH, CRYPTOS_LEDGER, ENDPOINT_DEFAULT, ENDPOINTS, ICON_DEFAULT, ICONS, LANGUAGE_DEFAULT, LEDGER_CONN, LEDGER_CONN_DEFAULT, LOCKING, LOCKING_DEFAULT, NOTIFICATION_DEFAULT, PREFIX_DEFAULT, PREFIXES, UIMODE_DEFAULT, UIMODES, UITHEME_DEFAULT, UITHEMES } from './defaults';
 
 type ChangeCallback = (settings: SettingsStruct) => void;
 type OnTypes = 'change';
@@ -24,6 +24,9 @@ function withDefault (options: Option[], option: string | undefined, fallback: s
 export class Settings implements SettingsStruct {
   readonly #emitter: EventEmitter;
 
+  #apiType: Endpoint;
+
+  // will become deprecated for supporting substrate connect light clients. apiType structure should be used instead
   #apiUrl: string;
 
   #camera: string;
@@ -42,12 +45,16 @@ export class Settings implements SettingsStruct {
 
   #uiTheme: string;
 
+  #notification: string;
+
   constructor () {
     const settings = (store.get('settings') as SettingsStruct) || {};
 
     this.#emitter = new EventEmitter();
 
-    this.#apiUrl = (typeof settings.apiUrl === 'string' && settings.apiUrl) || process.env.WS_URL || (ENDPOINT_DEFAULT.value as string);
+    // will become deprecated for supporting substrate connect light clients. apiType structure should be used instead
+    this.#apiUrl = (typeof settings.apiUrl === 'string' && settings.apiUrl) || (hasProcess && process.env && process.env.WS_URL) || (ENDPOINT_DEFAULT.value as string);
+    this.#apiType = { param: this.#apiUrl, type: 'json-rpc' as EndpointType };
     this.#camera = withDefault(CAMERA, settings.camera, CAMERA_DEFAULT);
     this.#ledgerConn = withDefault(LEDGER_CONN, settings.ledgerConn, LEDGER_CONN_DEFAULT);
     this.#i18nLang = settings.i18nLang || LANGUAGE_DEFAULT;
@@ -56,10 +63,15 @@ export class Settings implements SettingsStruct {
     this.#prefix = isUndefined(settings.prefix) ? PREFIX_DEFAULT : settings.prefix;
     this.#uiMode = settings.uiMode || UIMODE_DEFAULT;
     this.#uiTheme = settings.uiTheme || UITHEME_DEFAULT;
+    this.#notification = settings.notification || NOTIFICATION_DEFAULT;
   }
 
   public get camera (): string {
     return this.#camera;
+  }
+
+  public get apiType (): Endpoint {
+    return this.#apiType;
   }
 
   public get apiUrl (): string {
@@ -72,6 +84,10 @@ export class Settings implements SettingsStruct {
 
   public get icon (): string {
     return this.#icon;
+  }
+
+  public get notification (): string {
+    return this.#notification;
   }
 
   public get ledgerConn (): string {
@@ -140,12 +156,14 @@ export class Settings implements SettingsStruct {
 
   public get (): SettingsStruct {
     return {
+      apiType: this.#apiType,
       apiUrl: this.#apiUrl,
       camera: this.#camera,
       i18nLang: this.#i18nLang,
       icon: this.#icon,
       ledgerConn: this.#ledgerConn,
       locking: this.#locking,
+      notification: this.#notification,
       prefix: this.#prefix,
       uiMode: this.#uiMode,
       uiTheme: this.#uiTheme
@@ -153,12 +171,14 @@ export class Settings implements SettingsStruct {
   }
 
   public set (settings: Partial<SettingsStruct>): void {
+    this.#apiType = settings.apiType || this.#apiType;
     this.#apiUrl = settings.apiUrl || this.#apiUrl;
     this.#camera = settings.camera || this.#camera;
     this.#ledgerConn = settings.ledgerConn || this.#ledgerConn;
     this.#i18nLang = settings.i18nLang || this.#i18nLang;
     this.#icon = settings.icon || this.#icon;
     this.#locking = settings.locking || this.#locking;
+    this.#notification = settings.notification || this.#notification;
     this.#prefix = isUndefined(settings.prefix) ? this.#prefix : settings.prefix;
     this.#uiMode = settings.uiMode || this.#uiMode;
     this.#uiTheme = settings.uiTheme || this.#uiTheme;