+++ packages/ui-settings/src/types.ts
@@ -9,12 +9,14 @@ export type Option = {
 }
 
 export interface SettingsStruct {
+  apiType: Endpoint;
   apiUrl: string;
   camera: string;
   i18nLang: string;
   icon: string;
   ledgerConn: string;
   locking: string;
+  notification: string;
   prefix: number;
   uiMode: string;
   uiTheme: string;
@@ -28,3 +30,10 @@ export interface NetworkSpecsStruct {
   title: string;
   unit: string;
 }
+
+export interface Endpoint {
+  type: EndpointType;
+  param: string;
+}
+
+export type EndpointType = 'json-rpc' | 'substrate-connect';